timeText UI

+++ Assets/Scripts/GameManager.cs
@@ -19,7 +19,7 @@ public enum BattleState
 
 public class GameManager : MonoBehaviourPunCallbacks
 {
-    // ½Ì±ÛÅæ
+    // ì‹±ê¸€í†¤
     public static GameManager Instance
     {
         get
@@ -30,38 +30,38 @@ public class GameManager : MonoBehaviourPunCallbacks
         }
     }
     private static GameManager instance;
-    // board º¯¼ö
+    // board ë³€ìˆ˜
     [SerializeField] GameObject tiles;
     public Tile[,] board;
     const int boardRow = 9;
     const int boardCol = 5;
-    // Turn Ã¼Å© º¯¼ö
+    // Turn ì²´í¬ ë³€ìˆ˜
     int mainTurnNum = 6;
     int currentTurn = 1;
-    // ¼±ÅÃ ½Ã°£ º¯¼ö
+    // ì„ íƒ ì‹œê°„ ë³€ìˆ˜
     // int maxTime = 30;
     // int currentTime = 0;
 
-    // ÁÖ»çÀ§ ¼ö ÀúÀå º¯¼ö
+    // ì£¼ì‚¬ìœ„ ìˆ˜ ì €ì¥ ë³€ìˆ˜
     int diceNum;
-    // ÇöÀç °ÔÀÓ ÇÁ·Î¼¼½º »óÅÂ
+    // í˜„ì¬ ê²Œì„ í”„ë¡œì„¸ìŠ¤ ìƒíƒœ
     BattleState state;
-    // Player °´Ã¼µé ÀúÀå.
+    // Player ê°ì²´ë“¤ ì €ì¥.
     GameObject myPlayerObject;
     Player myPlayer;
-    // player »ı¼º
+    // player ìƒì„±
     [SerializeField] Transform[] spawnPositions;
     [SerializeField] GameObject playerPrefab;
     // processing flag
     bool isProcessing = false;
-    // obstacle ¼±ÅÃ flag
+    // obstacle ì„ íƒ flag
     bool isObstacleSelected;
-    // obstacle ¹öÆ°À» ¼±ÅÃÇÏµç move ¹öÆ°À» ¼±ÅÃÇÏµç ¼±ÅÃÇÏ¸é true
+    // obstacle ë²„íŠ¼ì„ ì„ íƒí•˜ë“  move ë²„íŠ¼ì„ ì„ íƒí•˜ë“  ì„ íƒí•˜ë©´ true
     bool isBtnSelected;
-    // Dice °´Ã¼ -> ÀÏ´Ü ¹Ì»ç¿ë
+    // Dice ê°ì²´ -> ì¼ë‹¨ ë¯¸ì‚¬ìš©
     //Dice dice;
 
-    // Time Ã¼Å© º¯¼ö 
+    // Time ì²´í¬ ë³€ìˆ˜ 
     [HideInInspector] public float currentTime = 0;
     float startTime = 0;
     [HideInInspector] public float maxTime = 30f;
@@ -70,7 +70,7 @@ public class GameManager : MonoBehaviourPunCallbacks
 
     void Start()
     {
-        // board ÃÊ±âÈ­
+        // board ì´ˆê¸°í™”
         board = new Tile[boardRow, boardCol];
         for (int i = 0; i < boardRow; i++)
         {
@@ -84,9 +84,9 @@ public class GameManager : MonoBehaviourPunCallbacks
                 Debug.Log("tile index : " + i + ", " + j);
             }
         }
-        // player »ı¼º
+        // player ìƒì„±
         SpanwPlayer();
-        // »óÅÂ ÃÊ±âÈ­
+        // ìƒíƒœ ì´ˆê¸°í™”
         state = BattleState.Start;
     }
 
@@ -104,27 +104,27 @@ public class GameManager : MonoBehaviourPunCallbacks
         GameProcess();
     }
 
-    // ½Ã°£ Ã¼Å©
+    // ì‹œê°„ ì²´í¬
     void CheckTime()
     {
-        // Àå¾Ö¹°°ú ÀÌµ¿ ¼±ÅÃ UI ¶ß´Â ¼ø°£ºÎÅÍ ½Ã°£ Àç±â ½ÃÀÛ
+        // ì¥ì• ë¬¼ê³¼ ì´ë™ ì„ íƒ UI ëœ¨ëŠ” ìˆœê°„ë¶€í„° ì‹œê°„ ì¬ê¸° ì‹œì‘
         currentTime = Time.time - startTime;
 
-        // ½Ã°£ ÁÙÀÌ¸é¼­ ÁÙÀÎ °ª UI¿¡ ¾÷µ¥ÀÌÆ®
+        // ì‹œê°„ ì¤„ì´ë©´ì„œ ì¤„ì¸ ê°’ UIì— ì—…ë°ì´íŠ¸
         Text txt = timeText.GetComponent<Text>();
-        txt.text = "" + (maxTime - currentTime);
+        txt.text = "" + (int)(maxTime - currentTime);
 
-        // ½Ã°£ÀÌ ´Ù Áö³ª¸é Time UI ºñÈ°¼ºÈ­
+        // ì‹œê°„ì´ ë‹¤ ì§€ë‚˜ë©´ Time UI ë¹„í™œì„±í™”
         if (currentTime >= maxTime)
         {
             timeText.SetActive(false);
             // network
             PhotonNetwork.LocalPlayer.SetCustomProperties(new Hashtable { { "isInputDone", true } });
-            // ½Ã°£ Ã¼Å© flag false·Î ¼³Á¤
+            // ì‹œê°„ ì²´í¬ flag falseë¡œ ì„¤ì •
             isTimeCheck = false;
         }
 
-        // ShowTimeUI, HideTimeUI() ±»ÀÌ ÇÊ¿ä ¾øÀ»µí
+        // ShowTimeUI, HideTimeUI() êµ³ì´ í•„ìš” ì—†ì„ë“¯
      
     }
 
@@ -141,27 +141,27 @@ public class GameManager : MonoBehaviourPunCallbacks
                 StartCoroutine(InputProcess());
                 break;
             case BattleState.SetObstacle:
-                // °ÔÀÓ¸Å´ÏÀúÀÇ Å¸ÀÏµéÀ» Å½»öÇØ¼­ Àå¾Ö¹° ÇÃ·¡±×°¡ ÀÖ´Â Å¸ÀÏ¿¡ Àå¾Ö¹° ½ÇÁ¦·Î ¼³Ä¡
+                // ê²Œì„ë§¤ë‹ˆì €ì˜ íƒ€ì¼ë“¤ì„ íƒìƒ‰í•´ì„œ ì¥ì• ë¬¼ í”Œë˜ê·¸ê°€ ìˆëŠ” íƒ€ì¼ì— ì¥ì• ë¬¼ ì‹¤ì œë¡œ ì„¤ì¹˜
                 if (isProcessing) return;
                 else isProcessing = true;
                 StartCoroutine(SetObstacle());
                 break;
             case BattleState.Move:
-                // ÀÌµ¿ ÀÔ·Â¿¡ µû¶ó¼­ ÇÃ·¹ÀÌ¾î¿¡¼­ ÀÌµ¿
-                // ¸¶Áö¸· ÅÏÀÌ¸é »óÅÂ Finish·Î º¯°æ
+                // ì´ë™ ì…ë ¥ì— ë”°ë¼ì„œ í”Œë ˆì´ì–´ì—ì„œ ì´ë™
+                // ë§ˆì§€ë§‰ í„´ì´ë©´ ìƒíƒœ Finishë¡œ ë³€ê²½
                 if (isProcessing) return;
                 else isProcessing = true;
                 StartCoroutine(AllMove());
                 break;
             case BattleState.Finish:
-                // ½ÂÆĞ ÆÇÁ¤
+                // ìŠ¹íŒ¨ íŒì •
                 if (isProcessing) return;
                 else isProcessing = true;
                 StartCoroutine(Finish());
                 break;
         }
     }
-    // ¸ğµç player°¡ ÀÔ·ÂÀ» ¹Ş¾Ò´ÂÁö Ã¼Å©
+    // ëª¨ë“  playerê°€ ì…ë ¥ì„ ë°›ì•˜ëŠ”ì§€ ì²´í¬
     private bool EveryPlayerReady()
     {
         foreach (Photon.Realtime.Player player in PhotonNetwork.PlayerList)
@@ -173,32 +173,32 @@ public class GameManager : MonoBehaviourPunCallbacks
         return true;
     }
 
-    // *Àå¾Ö¹° UI Ç¥½Ã
+    // *ì¥ì• ë¬¼ UI í‘œì‹œ
     private void ShowSelectUI()
     {
-        // Àå¾Ö¹°, ÀÌµ¿ ¼±ÅÃ UI ¸¸µé±â
-        // *Àå¾Ö¹°, ÀÌµ¿ ¼±ÅÃ UI Ç¥½Ã -> SetActive() ÇÔ¼ö ÀÌ¿ë
-        // Àå¾Ö¹° ¹öÆ° ¼±ÅÃ½Ã OnClick Listener -> OnClickObstacleBtn()
-        // ÀÌµ¿ ¹öÆ° ¼±ÅÃ½Ã OnClick Listener -> OnClickMoveBtn()
+        // ì¥ì• ë¬¼, ì´ë™ ì„ íƒ UI ë§Œë“¤ê¸°
+        // *ì¥ì• ë¬¼, ì´ë™ ì„ íƒ UI í‘œì‹œ -> SetActive() í•¨ìˆ˜ ì´ìš©
+        // ì¥ì• ë¬¼ ë²„íŠ¼ ì„ íƒì‹œ OnClick Listener -> OnClickObstacleBtn()
+        // ì´ë™ ë²„íŠ¼ ì„ íƒì‹œ OnClick Listener -> OnClickMoveBtn()
 
     }
-    // *ÁÖ»çÀ§ UI ¾÷µ¥ÀÌÆ®
+    // *ì£¼ì‚¬ìœ„ UI ì—…ë°ì´íŠ¸
     private void ChangeDiceUI()
     {
-        // dice ¼ö¸¦ Ç¥½ÃÇÏ´Â UI ¸¸µé±â
-        // *dice ¼ö¸¦ Ç¥½ÃÇÏ´Â UI¸¦ ¹Ş¾Æ¿Í¼­ ¾÷µ¥ÀÌÆ®
-        // NewGameMgr ÀÇ º¯¼ö diceNum ÀÌ¿ë
+        // dice ìˆ˜ë¥¼ í‘œì‹œí•˜ëŠ” UI ë§Œë“¤ê¸°
+        // *dice ìˆ˜ë¥¼ í‘œì‹œí•˜ëŠ” UIë¥¼ ë°›ì•„ì™€ì„œ ì—…ë°ì´íŠ¸
+        // NewGameMgr ì˜ ë³€ìˆ˜ diceNum ì´ìš©
     }
-    // Time ÁÙ¾îµå´Â ÇÔ¼ö
+    // Time ì¤„ì–´ë“œëŠ” í•¨ìˆ˜
     IEnumerator TimeCount()
     {
-        // *Time UI ¶ç¿ì±â
+        // *Time UI ë„ìš°ê¸°
         ShowTimeUI();
-        // *½Ã°£ ÁÙÀÌ¸é¼­ ÁÙÀÎ °ª UI¿¡ ¾÷µ¥ÀÌÆ®
-        // yield »ç¿ëÇØ¼­ ±¸Çö
-        // currentTime, maxTime º¯¼ö ÀÌ¿ë
-        // ½Ã°£ÀÌ ´Ù Áö³ª¸é Time UI ºñÈ°¼ºÈ­
-        // battle state ¸¦ setObstacle·Î º¯°æ
+        // *ì‹œê°„ ì¤„ì´ë©´ì„œ ì¤„ì¸ ê°’ UIì— ì—…ë°ì´íŠ¸
+        // yield ì‚¬ìš©í•´ì„œ êµ¬í˜„
+        // currentTime, maxTime ë³€ìˆ˜ ì´ìš©
+        // ì‹œê°„ì´ ë‹¤ ì§€ë‚˜ë©´ Time UI ë¹„í™œì„±í™”
+        // battle state ë¥¼ setObstacleë¡œ ë³€ê²½
         if (currentTime >= maxTime)
         {
 
@@ -206,34 +206,34 @@ public class GameManager : MonoBehaviourPunCallbacks
             PhotonNetwork.LocalPlayer.SetCustomProperties(new Hashtable { { "isInputDone", true } });
         }
 
-        // ¾Æ·¡ ÄÚµå´Â ¿À·ù ¹æÁö¸¦ À§ÇÑ ÄÚµå 
+        // ì•„ë˜ ì½”ë“œëŠ” ì˜¤ë¥˜ ë°©ì§€ë¥¼ ìœ„í•œ ì½”ë“œ 
         yield return null;
     }
-    // *Time UI ¶ç¿ì±â
+    // *Time UI ë„ìš°ê¸°
     void ShowTimeUI()
     {
         timeText.SetActive(true);
     }
-    // *Time UI ¼û±â±â
+    // *Time UI ìˆ¨ê¸°ê¸°
     void HideTimeUI()
     {
         timeText.SetActive(false);
     }
 
 
-    // dice Å¬·¡½º »ç¿ë¾ÈÇÒ½Ã
+    // dice í´ë˜ìŠ¤ ì‚¬ìš©ì•ˆí• ì‹œ
     void RollingDice()
     {
-        // ¸¶½ºÅÍ Å¬¶óÀÌ¾ğÆ®¿¡¼­¸¸ È£Ãâ
+        // ë§ˆìŠ¤í„° í´ë¼ì´ì–¸íŠ¸ì—ì„œë§Œ í˜¸ì¶œ
         if (PhotonNetwork.IsMasterClient)
         {
             diceNum = Random.Range(1, 7);
-            // dice ¾Ö´Ï¸ŞÀÌ¼Ç ÇÊ¿ä
+            // dice ì• ë‹ˆë©”ì´ì…˜ í•„ìš”
         }
     }
 
-    // RPC ÇÔ¼ö¸¦ ¿©±â¿¡¼­ Á¤ÀÇÇØµµ µÇ³ª
-    // ¾ÈµÈ´Ù..
+    // RPC í•¨ìˆ˜ë¥¼ ì—¬ê¸°ì—ì„œ ì •ì˜í•´ë„ ë˜ë‚˜
+    // ì•ˆëœë‹¤..
     //[PunRPC]
     //void SetDiceNum(int diceNum)
     //{
@@ -251,68 +251,68 @@ public class GameManager : MonoBehaviourPunCallbacks
     public void OnClickObstacleBtn()
     {
         isObstacleSelected = true;
-        // *¹öÆ° Å¬¸¯ flag (isBtnSelected) ÂüÀ¸·Î ¼³Á¤.
-        // *Btn UI ºñÈ°¼ºÈ­
+        // *ë²„íŠ¼ í´ë¦­ flag (isBtnSelected) ì°¸ìœ¼ë¡œ ì„¤ì •.
+        // *Btn UI ë¹„í™œì„±í™”
     }
     public void OnClickMoveBtn()
     {
         isObstacleSelected = false;
-        // *¹öÆ° Å¬¸¯ flag (isBtnSelected) ÂüÀ¸·Î ¼³Á¤.
-        // *Btn UI ºñÈ°¼ºÈ­
+        // *ë²„íŠ¼ í´ë¦­ flag (isBtnSelected) ì°¸ìœ¼ë¡œ ì„¤ì •.
+        // *Btn UI ë¹„í™œì„±í™”
     }
 
     // coroutine
     IEnumerator InputProcess()
     {
-        // ÁÖ»çÀ§ ±¼¸®±â -> master client¸¸ È£Ãâ
-        RollingDice();  // delay ÁÖ±â
+        // ì£¼ì‚¬ìœ„ êµ´ë¦¬ê¸° -> master clientë§Œ í˜¸ì¶œ
+        RollingDice();  // delay ì£¼ê¸°
         yield return new WaitForSeconds(2f);
 
-        // ÁÖ»çÀ§ µ¿±âÈ­
+        // ì£¼ì‚¬ìœ„ ë™ê¸°í™”
         SyscDiceNum();
 
-        // *ÁÖ»çÀ§ ´« ¼ö UI¿¡ Ç¥½Ã
+        // *ì£¼ì‚¬ìœ„ ëˆˆ ìˆ˜ UIì— í‘œì‹œ
         ChangeDiceUI();
 
-        // UI·Î Àå¾Ö¹° ³õÀ»°ÇÁö ÀÌµ¿ÇÒ°ÇÁö ÀÔ·Â ¹ŞÀ½
-        // *Àå¾Ö¹°, ÀÌµ¿ ¼±ÅÃ UI Ç¥½Ã
+        // UIë¡œ ì¥ì• ë¬¼ ë†“ì„ê±´ì§€ ì´ë™í• ê±´ì§€ ì…ë ¥ ë°›ìŒ
+        // *ì¥ì• ë¬¼, ì´ë™ ì„ íƒ UI í‘œì‹œ
         ShowSelectUI();
 
-        // *½Ã°£ Á¦ÇÑ ÇÔ¼ö(½Ã°£ count)
+        // *ì‹œê°„ ì œí•œ í•¨ìˆ˜(ì‹œê°„ count)
         // StartCoroutine(TimeCount());
-        // ½Ã°£ Àç´Â flag true ¼³Á¤
+        // ì‹œê°„ ì¬ëŠ” flag true ì„¤ì •
         isTimeCheck = true;
         startTime = Time.time;
         ShowTimeUI();
 
-        // ¹öÆ°ÀÌ ¼±ÅÃµÇÁö ¾ÊÀ¸¸é ´ë±â
-        // ÇÏÁö¸¸ ¹öÆ°ÀÌ ¼±ÅÃµÇÁö ¾Ê´õ¶óµµ ½Ã°£ÀÌ ÃÊ°úµÇ¸é ºüÁ®³ª¿À±â 
+        // ë²„íŠ¼ì´ ì„ íƒë˜ì§€ ì•Šìœ¼ë©´ ëŒ€ê¸°
+        // í•˜ì§€ë§Œ ë²„íŠ¼ì´ ì„ íƒë˜ì§€ ì•Šë”ë¼ë„ ì‹œê°„ì´ ì´ˆê³¼ë˜ë©´ ë¹ ì ¸ë‚˜ì˜¤ê¸° 
         while (!isBtnSelected && currentTime <= maxTime)
         {
             yield return null;
         }
-        // ÀÔ·Â¿¡ µû¶ó ÀÌµ¿, Àå¾Ö¹° ¼³Ä¡ ÀÔ·Â ÇÔ¼ö ½ÇÇà 
+        // ì…ë ¥ì— ë”°ë¼ ì´ë™, ì¥ì• ë¬¼ ì„¤ì¹˜ ì…ë ¥ í•¨ìˆ˜ ì‹¤í–‰ 
         if (isBtnSelected)
         {
             if (isObstacleSelected)
             {
-                // *ÀÌµ¿ ¼±ÅÃÇßÀ» °æ¿ì, ÇÃ·¹ÀÌ¾î¿¡¼­ ÀÌµ¿ ÀÔ·Â ¹ŞÀ½
+                // *ì´ë™ ì„ íƒí–ˆì„ ê²½ìš°, í”Œë ˆì´ì–´ì—ì„œ ì´ë™ ì…ë ¥ ë°›ìŒ
                 myPlayer.InputObstacle();
             }
             else
             {
-                // *Àå¾Ö¹° ¼³Ä¡ ¼±ÅÃÇßÀ» °æ¿ì, ÇÃ·¹ÀÌ¾î¿¡¼­ Àå¾Ö¹° ¼³Ä¡ ÀÔ·Â ¹ŞÀ½
+                // *ì¥ì• ë¬¼ ì„¤ì¹˜ ì„ íƒí–ˆì„ ê²½ìš°, í”Œë ˆì´ì–´ì—ì„œ ì¥ì• ë¬¼ ì„¤ì¹˜ ì…ë ¥ ë°›ìŒ
                 myPlayer.InputMove(diceNum);
             }
         }
 
-        // ³×Æ®¿öÅ©¿¡ ¸ğµç ÇÃ·¹ÀÌ¾î°¡ ÀÔ·ÂÀÌ ¿Ï·á µÆ´ÂÁö È®ÀÎ
+        // ë„¤íŠ¸ì›Œí¬ì— ëª¨ë“  í”Œë ˆì´ì–´ê°€ ì…ë ¥ì´ ì™„ë£Œ ëëŠ”ì§€ í™•ì¸
         // if (EveryPlayerReady())
         // {
         //     state = BattleState.SetObstacle;
         //     isProcessing = false;
         // }
-        // µÑ´Ù ¿Ï¼ºµÇÁö ¾ÊÀ¸¸é ´ë±â
+        // ë‘˜ë‹¤ ì™„ì„±ë˜ì§€ ì•Šìœ¼ë©´ ëŒ€ê¸°
         while (!EveryPlayerReady())
         {
             yield return null;
@@ -321,36 +321,36 @@ public class GameManager : MonoBehaviourPunCallbacks
         isProcessing = false;
     }
 
-    // *board¸¦ µÚÁ®¼­ obstacle ¼³Ä¡
+    // *boardë¥¼ ë’¤ì ¸ì„œ obstacle ì„¤ì¹˜
     IEnumerator SetObstacle()
     {
 
-        // *isObstacle flag°¡ »õ¿öÁ® ÀÖ´Â tileµéÀ» Ã£°í ±× tileµéÀÇ tileIndex °¡Á®¿À±â
-        // tileÀÇ index¸¦ ÅëÇØ Àå¾Ö¹° ¼³Ä¡
-        // Àå¾Ö¹°À» prefabÀ¸·Î ¹Ş¾Æ¼­ tileÀÇ tileIndexÀ» ÀÌ¿ëÇØ¼­ ÇØ´ç À§Ä¡¿¡ Àå¾Ö¹° spawn
+        // *isObstacle flagê°€ ìƒˆì›Œì ¸ ìˆëŠ” tileë“¤ì„ ì°¾ê³  ê·¸ tileë“¤ì˜ tileIndex ê°€ì ¸ì˜¤ê¸°
+        // tileì˜ indexë¥¼ í†µí•´ ì¥ì• ë¬¼ ì„¤ì¹˜
+        // ì¥ì• ë¬¼ì„ prefabìœ¼ë¡œ ë°›ì•„ì„œ tileì˜ tileIndexì„ ì´ìš©í•´ì„œ í•´ë‹¹ ìœ„ì¹˜ì— ì¥ì• ë¬¼ spawn
 
 
-        // 2ÃÊ ´ë±â
+        // 2ì´ˆ ëŒ€ê¸°
         yield return new WaitForSeconds(2f);
         state = BattleState.Move;
         isProcessing = false;
     }
 
-    // **¸ğµç ÇÃ·¹ÀÌ¾î ÀÌµ¿ÇÏ¸é¼­ Å¸ÀÏ »öÄ¥ 
+    // **ëª¨ë“  í”Œë ˆì´ì–´ ì´ë™í•˜ë©´ì„œ íƒ€ì¼ ìƒ‰ì¹  
     IEnumerator AllMove()
     {
-        // *myPlayer ¸â¹ö º¯¼ö¸¦ ÀÌ¿ëÇØ¼­ ÀÌµ¿ + »öÄ¥
-        // ÀÌµ¿ + »öÄ¥Àº RPC ÇÔ¼ö ¾È¿¡¼­ ±¸ÇöÇØ¾ßÇÒµí
-        // »öÄ¥½Ã ¿¹¿Ü Á¶°Ç ÇÊ¿ä
-        // AllMove()¸¦ ±¸ÇöÇÏ±â À§ÇØ¼­ Player¿¡ ÇÊ¿äÇÑ ÇÔ¼ö°¡ ÀÖÀ»½Ã µû·Î ¸¸µé±â
-        // ÀÌµ¿Àº playerÀÇ buffer¸¦ ÅëÇØ¼­ ÇØ´ç °æ·Î·Î ÀÌµ¿
-        // ¸¸¾à buffer°¡ ºñ¾î ÀÖÀ¸¸é (ÀÔ·Â°ªÀÌ ¾øÀ¸¸é) return
-        // ÀÌµ¿½Ã delay ÇÊ¿äÇÒµí -> yield¹® È°¿ë
-        // *ÀÌµ¿½Ã boardÀÇ tile¿¡ Àå¾Ö¹°ÀÌ ¼³Ä¡µÇ¾î ÀÖÀ¸¸é ÇØ´ç ÀÌµ¿ ºÒ°¡
-        // Àâ±â ½Ã½ºÅÛ ÇÊ¿ä -> ÀÏ´Ü Áö±İÀº ±¸Çö ¾ÈÇÏ´Â °É·Î (³×Æ®¿öÅ© ºÎºĞÀÌ µé¾î°¡¼­ ¾ê±âÇØºÁ¾ßÇÔ)
+        // *myPlayer ë©¤ë²„ ë³€ìˆ˜ë¥¼ ì´ìš©í•´ì„œ ì´ë™ + ìƒ‰ì¹ 
+        // ì´ë™ + ìƒ‰ì¹ ì€ RPC í•¨ìˆ˜ ì•ˆì—ì„œ êµ¬í˜„í•´ì•¼í• ë“¯
+        // ìƒ‰ì¹ ì‹œ ì˜ˆì™¸ ì¡°ê±´ í•„ìš”
+        // AllMove()ë¥¼ êµ¬í˜„í•˜ê¸° ìœ„í•´ì„œ Playerì— í•„ìš”í•œ í•¨ìˆ˜ê°€ ìˆì„ì‹œ ë”°ë¡œ ë§Œë“¤ê¸°
+        // ì´ë™ì€ playerì˜ bufferë¥¼ í†µí•´ì„œ í•´ë‹¹ ê²½ë¡œë¡œ ì´ë™
+        // ë§Œì•½ bufferê°€ ë¹„ì–´ ìˆìœ¼ë©´ (ì…ë ¥ê°’ì´ ì—†ìœ¼ë©´) return
+        // ì´ë™ì‹œ delay í•„ìš”í• ë“¯ -> yieldë¬¸ í™œìš©
+        // *ì´ë™ì‹œ boardì˜ tileì— ì¥ì• ë¬¼ì´ ì„¤ì¹˜ë˜ì–´ ìˆìœ¼ë©´ í•´ë‹¹ ì´ë™ ë¶ˆê°€
+        // ì¡ê¸° ì‹œìŠ¤í…œ í•„ìš” -> ì¼ë‹¨ ì§€ê¸ˆì€ êµ¬í˜„ ì•ˆí•˜ëŠ” ê±¸ë¡œ (ë„¤íŠ¸ì›Œí¬ ë¶€ë¶„ì´ ë“¤ì–´ê°€ì„œ ì–˜ê¸°í•´ë´ì•¼í•¨)
 
 
-        // Á¾·á Á¶°Ç È®ÀÎ
+        // ì¢…ë£Œ ì¡°ê±´ í™•ì¸
         currentTurn++;
         if (currentTurn > mainTurnNum) state = BattleState.Finish;
         else state = BattleState.Input;
@@ -358,12 +358,12 @@ public class GameManager : MonoBehaviourPunCallbacks
         yield return null;
     }
 
-    // **½ÂÆĞ È®ÀÎ
+    // **ìŠ¹íŒ¨ í™•ì¸
     IEnumerator Finish()
     {
 
-        // * board¸¦ Å½»öÇØ¼­ player1°ú player2ÀÇ ¿µ¿ª Ã£±â 
-        // * ½ÂÆĞ UI Ç¥½Ã
+        // * boardë¥¼ íƒìƒ‰í•´ì„œ player1ê³¼ player2ì˜ ì˜ì—­ ì°¾ê¸° 
+        // * ìŠ¹íŒ¨ UI í‘œì‹œ
         yield return null;
     }
